pyproject.toml
[project]
name = "example"
version = "0.1.0"
description = ""
authors = [
    {name = "Gonzalo Ayuso",email = "gonzalo123@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "langchain-aws (>=0.2.26,<0.3.0)",
    "click (>=8.2.1,<9.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
]

[tool.poetry]
name = "example"
version = "0.1.0"
description = ""
authors = ["Gonzalo Ayuso <gonzalo123@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
langchain-aws = ">=0.2.26,<0.3.0"
click = ">=8.2.1,<9.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
langchain = ">=0.3.26,<0.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
flake8 = "^7.3.0"
black = "^25.1.0"
isort = "^6.0.1"
pytest-watch = "^4.2.0"
mypy = "^1.8.0"
types-boto3 = "^1.34.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--tb=short",
    "--verbose",
    "--color=yes",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=85"
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
    "aws: Tests that require AWS services",
    "llm: Tests that use LLM services"
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
src_paths = ["src", "tests"]

[tool.flake8]
max-line-length = 88
extend-ignore = [
    "E203",  # whitespace before ':'
    "E501",  # line too long (handled by black)
    "W503",  # line break before binary operator
]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    ".eggs",
    "*.egg",
    "build",
    "dist",
]
per-file-ignores = [
    "__init__.py:F401",  # imported but unused
]

